refactor(workspace): move the record client out of the session package
Sessions, the namespace node table and workspace metadata are three
tables that persist the same way, so the keyed-record clients belong
under workspace/record/ rather than inside one of their consumers.
Before this the namespace and the state store imported upward into the
session package for a client that knows nothing about sessions.

A layering test in each language fails if the record tier grows an
import back into a tier that persists through it.
